completed exercises 5.1-5.10

+++ part4/bloglist/controllers/blogs.js
@@ -16,13 +16,19 @@ blogsRouter.get('/', async(request, response, next) => {
 blogsRouter.post('/', async(request, response, next) => {
 	try {
 		const body = request.body
-    const user = request.user
-		body["user"] = user._id
-		const blog = new Blog(body)
-		const savedBlog = await blog.save()
-		user.blogs = user.blogs.concat(savedBlog._id)
-		await user.save()
-		response.status(201).json(savedBlog)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    console.log(user)
+    //const user = request.user
+    body["user"] = user._id
+    const blog = new Blog(body)
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.status(201).json(savedBlog)
 	} catch (error) {
 		next(error)
 	}
@@ -31,9 +37,15 @@ blogsRouter.post('/', async(request, response, next) => {
 blogsRouter.delete('/:id', async(request, response, next) => {
 	try {
 		const blog = await Blog.findById(request.params.id)
-		const user = request.user
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+		//const user = request.user
 		if (blog.user.toString() === user._id.toString()) {
 			await Blog.findByIdAndRemove(blog.id)
+      
 		}
 		response.status(204).end()
 	} catch (error) {
